New package: xbindkeys-config
The patch is borrowed verbatim from Debian. It is under the same
license as the rest of the package, and is necessary for the app to
work.

There does not seem to be a canonical upstream homepage for this app
other than Debian itself.

Change-Id: I4f0bec32a95778fb12a48bfb280386b65a0d0f9a
(cherry picked from commit ea446390fad344509abe3ed2e740aa7dcc47d296)